Remove old default values for tank dimensions in towall and dwall
This removes the default values for tank dimensions in the utility functions calculating towall and dwall.

Now the tank dimensions must always be provided. So this change should prevent accidentally using the incorrect default values. The default values are not even valid for any detector any more (they were old IWCD values) and I don't think it makes sense to assume one particular detector by default, so it's better to enforce explicitly providing the dimensions since they are always required for the calculation.

This will break some old notebooks that relied on the default values, but that's intentional. If those notebooks are copied for new work, the dimensions of whatever detector would need to be provided in any case.